refactor: move repository interfaces to folder
