+++ Recademy.Application/Services/Implementations/ReviewService.cs
@@ -3,7 +3,7 @@ using Recademy.Core.Models.Projects;
 using Recademy.Core.Models.Reviews;
 using Recademy.Core.Tools;
 using Recademy.Core.Types;
-using Recademy.DataAccess.Repositories;
+using Recademy.DataAccess.Repositories.Abstractions;
 using Recademy.Shared.Dtos.Reviews;
 
 namespace Recademy.Application.Services.Implementations
