+++ Recademy.DataAccess/Repositories/Implementations/ProjectRepository.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore;
 using Recademy.Core.Models.Projects;
 using Recademy.Core.Models.Users;
 using Recademy.Core.Types;
+using Recademy.DataAccess.Repositories.Abstractions;
 
 namespace Recademy.DataAccess.Repositories.Implementations
 {
